requirements.txt
# Core Flask Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
Flask-Login==0.6.3

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Security & Authentication
Werkzeug==3.0.1
cryptography==41.0.7
bcrypt==4.1.2
PyJWT==2.8.0

# Environment & Configuration
python-dotenv==1.0.0

# HTTP & API
requests==2.31.0

# Payment Processing
stripe==7.8.0

# Server
gunicorn==21.2.0
gevent==23.9.1

# Geolocation & Maps
geopy==2.4.1
geoip2==4.7.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Image Processing
Pillow==10.1.0
qrcode==7.4.2

# Machine Learning
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
python-dateutil==2.8.2
user-agents==2.2.0
redis==5.0.1
psutil==5.9.6

# Telegram Bot
python-telegram-bot==20.7

# Rate Limiting
Flask-Limiter==3.5.0

# Input Sanitization
bleach==6.1.0

# Scheduling (for cron jobs)
APScheduler==3.10.4
